+++ README.md
@@ -60,20 +60,20 @@ yarn install
 ### Exécution
 
 - `yarn test` – lance l'ensemble de la suite de tests.
-- `yarn test:unit` – exécute les tests unitaires (`test/unit`).
-- `yarn test:api` – exécute les tests d'API (`test/api`).
-- `yarn test:e2e` – exécute les tests end-to-end avec Playwright (`test/e2e`).
+- `yarn test:unit` – exécute les tests unitaires (`tests/unit`).
+- `yarn test:api` – exécute les tests d'API (`tests/api`).
+- `yarn test:e2e` – exécute les tests end-to-end avec Playwright (`tests/e2e`).
 
 ### Ajouter un test
 
 1. Créer le fichier dans le dossier approprié :
-    - `test/unit` pour les tests unitaires,
-    - `test/api` pour les tests d'API,
-    - `test/e2e` pour les tests end-to-end.
+    - `tests/unit` pour les tests unitaires,
+    - `tests/api` pour les tests d'API,
+    - `tests/e2e` pour les tests end-to-end.
 2. Nommer les fichiers selon la convention `*.test.ts`.
-3. Les anciens tests sont sauvegardés dans `test/_legacy`.
+3. Les anciens tests sont sauvegardés dans `tests/_legacy`.
 
 ### Mock d'AWS Amplify
 
-Le mock d'AWS Amplify et la configuration partagée se trouvent dans `test/setupTests.ts`.
+Le mock d'AWS Amplify et la configuration partagée se trouvent dans `tests/_legacy/setupTests.ts`.
 Ce fichier configure Amplify et le serveur MSW avant chaque test.
