+++ playwright.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig, devices } from "playwright/test";
 
 export default defineConfig({
-    testDir: "./e2e",
+    testDir: "./tests/e2e",
     reporter: [["html", { open: "never" }]],
     use: {
         baseURL: process.env.PLAYWRIGHT_BASE_URL || "http://localhost:3000",
