+++ tests/e2e/menu.spec.ts
