+++ tests/e2e/section-management.spec.ts
