+++ tests/e2e/setupPaths.ts
