+++ tsconfig.json
@@ -1,7 +1,11 @@
 {
     "compilerOptions": {
         "target": "ES2017",
-        "lib": ["dom", "dom.iterable", "esnext"],
+        "lib": [
+            "dom",
+            "dom.iterable",
+            "esnext"
+        ],
         "allowJs": true,
         "skipLibCheck": true,
         "strict": true,
@@ -17,20 +21,48 @@
         "incremental": true,
         "baseUrl": ".",
         "paths": {
-            "@/*": ["./*"],
-            "@app/*": ["app/*"],
-            "@src/*": ["src/*"],
-            "@amplify/*": ["amplify/*"],
-            "@components/*": ["src/components/*"],
-            "@hooks/*": ["src/hooks/*"],
-            "@context/*": ["src/context/*"],
-            "@utils/*": ["src/utils/*"],
-            "@assets/*": ["src/assets/*"],
-            "@services/*": ["src/services/*"],
-            "@myTypes/*": ["src/types/*"],
-            "@entities/*": ["src/entities/*"],
-            "@public/*": ["public/*"],
-            "@test/*": ["test/*"]
+            "@/*": [
+                "./*"
+            ],
+            "@app/*": [
+                "app/*"
+            ],
+            "@src/*": [
+                "src/*"
+            ],
+            "@amplify/*": [
+                "amplify/*"
+            ],
+            "@components/*": [
+                "src/components/*"
+            ],
+            "@hooks/*": [
+                "src/hooks/*"
+            ],
+            "@context/*": [
+                "src/context/*"
+            ],
+            "@utils/*": [
+                "src/utils/*"
+            ],
+            "@assets/*": [
+                "src/assets/*"
+            ],
+            "@services/*": [
+                "src/services/*"
+            ],
+            "@myTypes/*": [
+                "src/types/*"
+            ],
+            "@entities/*": [
+                "src/entities/*"
+            ],
+            "@public/*": [
+                "public/*"
+            ],
+            "@test/*": [
+                "tests/*"
+            ]
         },
         "plugins": [
             {
@@ -39,7 +71,6 @@
         ],
         "strictNullChecks": true
     },
-
     "include": [
         "next-env.d.ts",
         "app/**/*.ts",
